break Pairing view out into more files
