+++ frontend/src/views/Pairing/index.js
@@ -1,7 +1,6 @@
 import React, { Fragment, useState } from "react";
 import Helmet from 'react-helmet';
 import { makeStyles } from '@material-ui/core/styles';
-import Button from '@material-ui/core/Button';
 import Accordion from '@material-ui/core/Accordion';
 import AccordionSummary from '@material-ui/core/AccordionSummary';
 import AccordionDetails from '@material-ui/core/AccordionDetails';
@@ -10,106 +9,16 @@ import DialogTitle from '@material-ui/core/DialogTitle';
 import DialogContent from '@material-ui/core/DialogContent';
 import DialogActions from '@material-ui/core/DialogActions';
 import DialogContentText from '@material-ui/core/DialogContentText';
-import Drawer from '@material-ui/core/Drawer';
-import Radio from '@material-ui/core/Radio';
-import Snackbar from '@material-ui/core/Snackbar';
 import IconButton from '@material-ui/core/IconButton';
 import CloseIcon from 'mdi-react/CloseIcon';
 import ChevronDownIcon from 'mdi-react/ChevronDownIcon';
-import Loading from '../components/Loading';
-import Participant from '../components/Participant';
-import LinearProgress from '@material-ui/core/LinearProgress';
-import { useParticipants } from '../data/participants';
-import { usePostPairing } from '../data/pairing';
+import Loading from '../../components/Loading';
+import ParticipantColumn from './ParticipantColumn';
+import PairDrawer from './PairDrawer';
+import PairedSnackbar from './PairedSnackbar';
+import { useParticipants } from '../../data/participants';
+import { usePostPairing } from '../../data/pairing';
 
-const useStyles = makeStyles(theme => ({
-  // participant columns are floated to achieve similar heights
-  participantColumn: {
-    float: 'left',
-    width: '50%',
-    padding: theme.spacing(1),
-  },
-  drawer: {
-    padding: theme.spacing(1),
-  },
-}));
-
-function ParticipantColumn({ participants, title, filter, onSelect, selected }) {
-  const classes = useStyles();
-
-  return (
-    <div className={classes.participantColumn}>
-      <h2>{title}</h2>
-      {participants.data
-        .filter(filter)
-        .map(p => (
-          <Accordion defaultExpanded key={p.id}>
-            <AccordionSummary expandIcon={<ChevronDownIcon />} aria-label="Expand">
-							<FormControlLabel
-								aria-label="Select"
-								onClick={(event) => { event.stopPropagation(); onSelect(p); }}
-								onFocus={(event) => event.stopPropagation()}
-								control={<Radio checked={Boolean(selected && selected.id === p.id)} />}
-								label={p.full_name} />
-            </AccordionSummary>
-            <AccordionDetails>
-              <Participant key={p.id} participant={p} />
-            </AccordionDetails>
-          </Accordion>
-        ))}
-    </div>
-  );
-}
-
-function PairDrawer({ mentor, learner, open, pairing, onClose, onPair }) {
-  const classes = useStyles();
-
-  // TODO: show time compatibility, shared interests, etc.
-
-  return (
-    <Drawer
-      classes={{ paper: classes.drawer }}
-      anchor="bottom"
-      open={open}
-      onClose={onClose}
-      variant="persistent"
-      PaperProps={{ elevation: 4 }}>
-      {open && (
-        <Fragment>
-          {pairing && <LinearProgress />}
-          <DialogTitle>Proposed Pair: {mentor.full_name} / {learner.full_name}</DialogTitle>
-          <DialogContent>
-            <DialogContentText>
-              This seems like a great match, nice job!
-            </DialogContentText>
-          </DialogContent>
-          <DialogActions>
-            <Button onClick={onClose} color="primary">
-              Cancel
-            </Button>
-            <Button onClick={() => onPair(mentor, learner)} color="primary" autoFocus>
-              Pair
-            </Button>
-          </DialogActions>
-        </Fragment>
-      )}
-    </Drawer>
-  );
-}
-
-function PairedSnackbar({ onClose, open, pair }) {
-  return (
-    <Snackbar
-      anchorOrigin={{
-        vertical: 'bottom',
-        horizontal: 'center',
-      }}
-      open={open}
-      autoHideDuration={10000}
-      onClose={onClose}
-      message={`Pair created: ${pair.mentor?.full_name} / ${pair.learner?.full_name}`} />
-  );
-}
 
 export default function Home(props) {
   const [participants, refetchParticipants] = useParticipants();
